Add relative cost measure P(x)/P(y)
When option x isn't chosen we measure the amount of "social injustice"
Even though x is preferred to y by P(x)/P(x) - it was still not
chosen. (P - can be any scoring rule or the number of voters in the
majority rule).
Since P(y) can be 0, we clip it to 0.1